--- server_side/text_editor/include/ISearchForPlace.hpp
@@ -1,10 +1,0 @@
-#include <string>
-
-#include "element.hpp"
-
-class IWorkWithText;
-
-class ISearchForPlace {
-    public:
-    virtual std::string findWhereToPlace(Element* elementToPlace) = 0;
-};
