+++ server_side/text_editor/include/SearcherForPlace.hpp
@@ -1,11 +1,16 @@
+#pragma once
 #include "ISearchForPlace.hpp"
 #include "TextEditor.hpp"
 
+
+bool checkForAfterElement(Element* afterElement, Element* checkElement);
+bool checkForBeforeLess(Element* elementBefore, Element* elementAfter);
+
+
 class Searcher : public ISearchForPlace {
     public:
-    Searcher(const WorkWithLines* whereToPlace);
-    std::string findWhereToPlace(Element* elementToPlace) override;
-    
-    private:
-    IWorkWithText* whereToPlace;
+    AnswerForInsertAction whatPositionDelete(Element* elementToDelete, Element* beggining, StartOfLine* lines) override;
+    AnswerForInsertAction whatPosition(Element* beforeInsert, Element* insertElement, Element* afterElement, StartOfLine* line) override;
+
+    Element* getToPosition(Element* fromWhere, size_t position) override;
 };
