+++ server_side/text_editor/include/TranslatorFromList.hpp
@@ -2,5 +2,6 @@
 #include "ITranslator.hpp"
 
 class TranslatorFromList : public ITranslator {
+    public:
     std::string returnStringFromDataType(Element* from) override;
 };
