+++ server_side/text_editor/include/WorkWithData.hpp
@@ -4,7 +4,7 @@
 #include "WorkWithLogFile.hpp"
 
 #include "TextEditor.hpp"
-// #include "TranslatorFromList.hpp"
+#include "TranslatorFromList.hpp"
 
 class WorkWithData : public IWorkWithData {
     public:
@@ -20,7 +20,7 @@ class WorkWithData : public IWorkWithData {
     std::string getFileWithDataDricetory() override;
 
     private:
-    // ITranslator* translator;
+    ITranslator* translator;
     IWorkWithFile* workWithCPP;
     IWorkWithFile* workWithLog;
