+++ server_side/text_editor/interfaces/ITextEditor.hpp
@@ -22,7 +22,7 @@ class IWorkWithText {
 
     virtual size_t getQuantityOfLines() = 0;
     virtual Element* getStartOfLine(size_t lineNumber) = 0;
-    virtual size_t getSizeOfLine(size_t lineNumber) = 0;
+    // virtual size_t getSizeOfLine(size_t lineNumber) = 0;
 
-    virtual AnswerLinePos getLinePosFromPos(size_t pos) = 0;
+    // virtual AnswerLinePos getLinePosFromPos(size_t pos) = 0;
 };
